changes in footer page

+++ app/Http/Controllers/Client/NomenclatureController.php
@@ -17,25 +17,27 @@ class NomenclatureController extends BaseController
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request){
-        $this->validate($request, [
-          'names.0' => 'required|string',
-        ]);
-        NomenclatureTranslation::truncate();
-        $language_ids = $request->language_ids;
-        $loyalty_cards_language_ids = $request->loyalty_cards_language_ids;
-        foreach ($request->names as $key => $name) {
-            if($name){
-                $nomenclature = NomenClature::where('label', 'vendors')->first();
-                if($nomenclature){
-                    $nomenclature->update(['label' => "vendors"]);
-                }else{
-                    $nomenclature = NomenClature::create(['label' => "vendors"]);
+        if(count($request->names) > 0){
+            $this->validate($request, [
+              'names.0' => 'required|string',
+            ]);
+            NomenclatureTranslation::truncate();
+            $language_ids = $request->language_ids;
+            $loyalty_cards_language_ids = $request->loyalty_cards_language_ids;
+            foreach ($request->names as $key => $name) {
+                if($name){
+                    $nomenclature = NomenClature::where('label', 'vendors')->first();
+                    if($nomenclature){
+                        $nomenclature->update(['label' => "vendors"]);
+                    }else{
+                        $nomenclature = NomenClature::create(['label' => "vendors"]);
+                    }
+                    $nomenclature_translation =  new NomenclatureTranslation();
+                    $nomenclature_translation->name = $name;
+                    $nomenclature_translation->language_id = $language_ids[$key];
+                    $nomenclature_translation->nomenclature_id = $nomenclature->id;
+                    $nomenclature_translation->save();
                 }
-                $nomenclature_translation =  new NomenclatureTranslation();
-                $nomenclature_translation->name = $name;
-                $nomenclature_translation->language_id = $language_ids[$key];
-                $nomenclature_translation->nomenclature_id = $nomenclature->id;
-                $nomenclature_translation->save();
             }
         }
         if(count($request->loyalty_cards_names) > 0){
